Implement package internal/configflags
- create package internal/configflags
- move flags setting logics to internal/configflags
